+++ src/managers/InfoManager.ts
@@ -22,7 +22,7 @@ export default class InfoManager extends BaseManager<Webhook> {
 			}),
 			headers: { "content-type": "application/json" },
 		}).then(w=>w.json())
-		if (add.error) throw new GenericAPIError(`${add.error}: ${add.description}`)
+		if (add.error) throw new GenericAPIError(`${add.error}: ${add.message}`)
 		return add
 	} 
 	async removeWebhook(webhookid: string, webhooktoken: string, guildid: string): Promise<Webhook> {