+++ src/managers/ReportManager.ts
@@ -23,14 +23,14 @@ export default class ReportManager extends BaseManager<Report> {
 		const fetched = await fetch(`${this.apiurl}/reports/getbyid?id=${strictUriEncode(reportid)}`).then(c=>c.json())
 
 		if (!fetched) return null // return null if the fetch is empty
-		if (fetched.error) throw new GenericAPIError(`${fetched.error}: ${fetched.description}`)
+		if (fetched.error) throw new GenericAPIError(`${fetched.error}: ${fetched.message}`)
 		if (cache) this.add(fetched)
 		return fetched
 	}
 	async fetchAllName(playername: string, cache=true): Promise<Report[]> {
 		const allReports = await fetch(`${this.apiurl}/reports/getall?playername=${strictUriEncode(playername)}`).then(c=>c.json())
 
-		if (allReports.error) throw new GenericAPIError(`${allReports.error}: ${allReports.description}`)
+		if (allReports.error) throw new GenericAPIError(`${allReports.error}: ${allReports.message}`)
 
 		if (cache && allReports[0]) {
 			allReports.forEach(report => this.add(report))
@@ -47,18 +47,19 @@ export default class ReportManager extends BaseManager<Report> {
 		if (cache) ruleReports.forEach(report => this.add(report))
 		return ruleReports
 	}
-	async create(report: CreateReport, cache = true, reqConfig: RequestConfig = {}): Promise<Report> {
+	async create(report: any, cache = true, reqConfig: RequestConfig = {}): Promise<Report> {
 		if (!this.apikey && !reqConfig.apikey) throw new NoApikeyError()
 
-		const create = await fetch(`${this.apiurl}/reports/create`, {
+		const create = await fetch(`${this.apiurl}/reports`, {
 			method: "POST",
 			body: JSON.stringify(report),
-			headers: { "apikey": this.apikey || reqConfig.apikey, "content-type": "application/json" },
+			headers: { "authorization": `Token ${this.apikey || reqConfig.apikey}`, "content-type": "application/json" },
 		}).then(u=>u.json())
 
 		if (create.error) {
 			if (create.description === "API key is wrong") throw new AuthenticationError()
-			throw new GenericAPIError(`${create.error}: ${create.description}`)
+			console.error(create)
+			throw new GenericAPIError(`${create.error}: ${create.message}`)
 		}
 		if (cache) this.add(create)
 		return create
@@ -70,12 +71,12 @@ export default class ReportManager extends BaseManager<Report> {
 				id: reportid,
 				adminId: adminId,
 			}),
-			headers: { "apikey": this.apikey || reqConfig.apikey, "content-type": "application/json" },
+			headers: { "authorization": `Token ${this.apikey || reqConfig.apikey}`, "content-type": "application/json" },
 		}).then(u=>u.json())
 
 		if (revoked.error) {
 			if (revoked.description === "API key is wrong") throw new AuthenticationError()
-			throw new GenericAPIError(`${revoked.error}: ${revoked.description}`)
+			throw new GenericAPIError(`${revoked.error}: ${revoked.message}`)
 		}
 
 		if (!revoked?.revokedTime) throw new UnsuccessfulRevocationError()
@@ -95,7 +96,7 @@ export default class ReportManager extends BaseManager<Report> {
 
 		if (revoked.error) {
 			if (revoked.description === "API key is wrong") throw new AuthenticationError()
-			throw new GenericAPIError(`${revoked.error}: ${revoked.description}`)
+			throw new GenericAPIError(`${revoked.error}: ${revoked.message}`)
 		}
 
 		revoked.forEach(revocation => {